Split metadata tests into own file as suggested. Added support for FormDef, MdsolHelp, ItemGroupRef. Fixed bug in ItemGroupRef Repeat (rwslib docs are wrong..sigh). Updated ODM root and allowed transaction type to be set to snapshot for MetaData
